Added a getHashtagList method.
